requirements.txt
pandas
numpy
PyMuPDF
google-cloud-vision
Pillow
